+++ qt project/NievaTrains_P2/model/gerarchia/bimode.h
@@ -12,19 +12,19 @@ private:
 
 public:
 
-    Bimode(const std::string& = "NoName", const std::string& ="NoBuilder", unsigned int =100, Trotaia =Trotaia::maglev, Ttreno =Ttreno::alta_velocita, TtrasmissioneElettrico =TtrasmissioneElettrico::overhead_lines, float =0.7f, float =0.4f, Tfuel =Tfuel::kerosene, TtrasmissioneFuel =TtrasmissioneFuel::electric, Tmotore =Tmotore::electric);
+    Bimode(const std::string& = "NoName", const std::string& ="NoBuilder", unsigned int =100, Trotaia =Trotaia::maglev, Ttreno =Ttreno::alta_velocita, TtrasmissioneElettrico =TtrasmissioneElettrico::overhead_lines, double =0.7, double =0.4, Tfuel =Tfuel::kerosene, TtrasmissioneFuel =TtrasmissioneFuel::electric, Tmotore =Tmotore::electric);
     std::string getMotorePrimario() const;
 
     void setMotorePrimario(std::string);
     void setCarburanteIC(std::string);
     void setTrasmissioneIC(std::string);
-    void setEfficenzaIC(float);
+    void setEfficenzaIC(double);
     void setTrasmissioneElettrico(std::string);
-    void setEfficenzaElettrico(float);
+    void setEfficenzaElettrico(double);
 
 
 
-    float carburanteNecessario(unsigned int) const;  //input km da percorrere
+    double carburanteNecessario(unsigned int) const;  //input km da percorrere
     unsigned int kmPercorribili(unsigned int) const; //input kw o kg di carburante(in base al motore primario)(se hybrid e' fuel kg)
     std::string type() const;
     void print() const;
